Return long instead of double for fitness

+++ com-mmxlabs-optimiser/optimiser-lso/src/main/java/com/mmxlabs/optimiser/lso/impl/LinearSimulatedAnnealingFitnessEvaluator.java
@@ -228,7 +228,7 @@ public final class LinearSimulatedAnnealingFitnessEvaluator<T> implements
 	 * 
 	 * @return
 	 */
-	public double getCurrentFitness() {
+	public long getCurrentFitness() {
 		return currentFitness;
 	}
 }
